+++ dsl/src/main/groovy/com/readytalk/jenkins/model/meta/AggregateField.groovy
@@ -3,10 +3,57 @@ package com.readytalk.jenkins.model.meta
 import com.readytalk.jenkins.model.ContextMap
 import com.readytalk.jenkins.model.ContextAlreadyBoundException
 import com.readytalk.jenkins.model.ItemSource
+import org.codehaus.groovy.runtime.metaclass.MethodSelectionException
+import sun.reflect.generics.reflectiveObjects.NotImplementedException
 
 abstract class AggregateField extends AbstractComponentAdapter {
   abstract String getAggregateField()
+
   boolean shouldInherit(ItemSource item) { true }
+
+  def combine(outer, inner) {
+    boolean noDefault = false
+
+    try {
+      noDefault = !(outer.respondsTo('plus', inner.class))
+    } catch(MethodSelectionException e) {
+      //Ignore - if we hit this, it'll still work via dynamic dispatch
+    }
+
+    if(!noDefault) {
+      return outer.plus(inner)
+    } else {
+      throw new UnsupportedOperationException(
+              "Cannot automatically combine ${aggregateField} values for types:" +
+              "${outer.class} + ${inner.class}\n" +
+              "Explicit override of AggregateField.combine(outer,inner) required."
+      )
+    }
+  }
+
+  def getUnit() {
+    def globalDefault = getFields().get(aggregateField)
+    switch(globalDefault) {
+      case String:
+        return ''
+      case Map:
+        return [:]
+      case Iterable:
+        return []
+      case Number:
+        return 0
+      case null:
+        throw new IllegalStateException(
+                "Global default for aggregate fields cannot be null!\n" +
+                "Component: ${getName()}, Field: ${aggregateField}"
+        )
+      default:
+        throw new UnsupportedOperationException(
+                "Unit value for type ${globalDefault.class.name}\n"+
+                        "Explicit getUnit() implementation required"
+       )
+    }
+  }
 
   //Aggregate values from all visible scopes for the given field in the implementing component
   //NOTE: This method is required to be idempotent
@@ -16,21 +63,23 @@ abstract class AggregateField extends AbstractComponentAdapter {
       String field = getAggregateField()
 
       //TODO: Should we block upward aggregation if an intermediate context has inherit = false?
+      //TODO: This probably won't work with templating
       Closure aggregate = { ContextMap context ->
         def iter = context
-        def map = [:]
+        def sum = getUnit()
         while(iter != null && iter instanceof ContextMap) {
-          map = (iter.context.get(componentName)?.get(field) ?: [:]) + map
+          sum = combine((iter.context.get(componentName)?.get(field) ?: getUnit()), sum)
           iter = iter.parent
         }
-        return map
+        return sum
       }
-      def defaultsMap = aggregate(item.defaults)
-      def userMap = aggregate(item.user)
 
-      def localScope = item.lookupValue(componentName, field) ?: [:]
+      def defaultsSum = aggregate(item.defaults)
+      def userSum = aggregate(item.user)
+
+      def localScope = item.lookupValue(componentName, field) ?: getUnit()
       try {
-        item.user.bind(componentName, field, defaultsMap + userMap + localScope)
+        item.user.bind(componentName, field, combine(combine(defaultsSum, userSum), localScope))
       } catch(ContextAlreadyBoundException e) {
         //Ignore - the point of this adapter specifically requires we bend the scoping rules a bit
       }