+++ dsl/src/main/groovy/com/readytalk/jenkins/model/types/CommonComponent.groovy
@@ -2,23 +2,41 @@ package com.readytalk.jenkins.model.types
 
 import com.readytalk.jenkins.model.AbstractComponentType
 import com.readytalk.jenkins.model.Fixed
+import com.readytalk.jenkins.model.ItemSource
+import com.readytalk.jenkins.model.meta.AggregateField
+import com.readytalk.jenkins.model.meta.ComponentAdapter
+import com.readytalk.util.StringUtils
 
 @Fixed
 class CommonComponent extends AbstractComponentType {
   String name = 'common'
+
+  List<ComponentAdapter> traits = [
+          new AggregateField() {
+            String aggregateField = 'description'
+
+            boolean shouldInherit(ItemSource item) {
+              return item.lookupValue(name, 'inheritDescription')
+            }
+
+            def combine(outer,inner) {
+              outer + (!(outer==''||inner=='') ? '\n' : '') + inner
+            }
+          }
+  ]
 
   Map<String, ?> fields = [
           runSchedule:      "@weekly",
           historyDays:      '',  //days to keep build history
           historyCount:     30,  //builds to keep if days not set
           description:      '',  //job description
-          //TODO: This should use the env,VAR syntax instead
           buildName:        '#${BUILD_NUMBER}-${GIT_BRANCH}',
           buildHost:        '',
           jdkVersion:       '(Default)',
           useXvnc:          false,
           injectPasswords:  false, //if true, injects global passwords into the build
           concurrentBuild:  false,
+          inheritDescription: true,
   ]
 
   Closure dslConfig = { vars->